+++ src/main/kotlin/arguments/MhguHzv.kt
@@ -3,9 +3,7 @@ package arguments
 import autocomplete.MhguMonsterAutoComplete
 import com.kotlindiscord.kord.extensions.commands.Arguments
 import com.kotlindiscord.kord.extensions.commands.converters.impl.string
-import dev.kord.common.annotation.KordPreview
 
-@KordPreview
 class MhguHzv: Arguments() {
     val monsterName by string {
         name = "monster_name"