+++ src/main/kotlin/autocomplete/MhwItem.kt
@@ -1,5 +1,6 @@
 package autocomplete
 
+import ApiClient
 import com.kotlindiscord.kord.extensions.utils.FilterStrategy
 import com.kotlindiscord.kord.extensions.utils.suggestStringMap
 import dev.kord.core.entity.interaction.AutoCompleteInteraction
@@ -15,6 +16,13 @@ val MhwItemAutoComplete: suspend AutoCompleteInteraction.(AutoCompleteInteractio
                 matchEntryKey = { it.key },
             )
         )
+
+        suggestStringMap(
+            results.associate { item -> Pair(item.value.name, item.value.name) },
+            FilterStrategy { _, _ -> true }
+        )
+    } else {
+        val results = ApiClient.MHW.items.entries.take(25)
 
         suggestStringMap(
             results.associate { item -> Pair(item.value.name, item.value.name) },