+++ src/main/kotlin/extensions/changelog.kt
@@ -4,16 +4,12 @@ import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.Color
-import dev.kord.common.annotation.KordPreview
-import dev.kord.common.entity.Snowflake
 import dev.kord.rest.builder.message.create.embed
 import kotlinx.datetime.Clock
 import utils.CatBot
 import utils.version
 import java.util.logging.Logger
 
-
-@OptIn(KordPreview::class)
 class Changelog: Extension() {
 
     companion object {
@@ -23,7 +19,10 @@ class Changelog: Extension() {
     override val name = "ChangeLog"
 
     val changelog = """
-        🆕 Rewrote Whole Ass Bot - (message commands to slash commands)
+        - General maintenance 
+        - Added MH rise category along with a HZV command.
+        - Changed error response to be an embed.
+        - Auto complete will now respond with first 25 results if empty string.
     """.trimIndent()
 
     override suspend fun setup() {
@@ -44,7 +43,7 @@ class Changelog: Extension() {
                         }
 
                         timestamp = Clock.System.now()
-                        footer { text = "Changelog Menu"; icon = kord.getSelf().avatar?.url }
+                        footer { text = "Changelog Menu"; icon = kord.getSelf().avatar?.cdnUrl?.toUrl() }
                     }
                 }
             }