+++ src/main/kotlin/extensions/invite.kt
@@ -5,8 +5,6 @@ import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.Color
-import dev.kord.common.annotation.KordPreview
-import dev.kord.common.entity.Snowflake
 import dev.kord.core.Kord
 import dev.kord.rest.builder.message.create.embed
 import kotlinx.coroutines.flow.first
@@ -14,7 +12,6 @@ import kotlinx.datetime.Clock
 import utils.CatBot
 import java.util.logging.Logger
 
-@OptIn(KordPreview::class)
 class Invite: Extension() {
     companion object {
         val LOG = Logger.getLogger(Invite::class.java.name)
@@ -43,7 +40,7 @@ class Invite: Extension() {
 
                         footer {
                             text = "Invite Link Request"
-                            icon = botUser?.avatar?.url
+                            icon = botUser?.avatar?.cdnUrl?.toUrl()
                         }
 
                         timestamp = Clock.System.now()