+++ src/main/kotlin/extensions/mhgu/Hzv.kt
@@ -2,31 +2,22 @@ package extensions.mhgu
 
 import arguments.MhguHzv
 import com.kotlindiscord.kord.extensions.commands.application.slash.PublicSlashCommand
+import com.kotlindiscord.kord.extensions.components.forms.ModalForm
 import com.kotlindiscord.kord.extensions.types.respond
 import com.kotlindiscord.kord.extensions.utils.capitalizeWords
 import dev.kord.common.Color
-import dev.kord.common.annotation.KordPreview
 import dev.kord.rest.NamedFile
 import dev.kord.rest.builder.message.create.embed
+import io.ktor.client.request.forms.*
+import io.ktor.utils.io.jvm.javaio.*
 import utils.CatBot
+import utils.capitalizeDashes
 import java.io.File
-import java.util.*
 import java.util.logging.Logger
 
-// this is so we can handle cases like Ahtal-Ka
-fun String.capitalizeDashes(locale: Locale? = null): String {
-    return split("-").joinToString("-") { word ->
-        word.replaceFirstChar {
-            if (it.isLowerCase()) {
-                it.titlecase(locale ?: Locale.getDefault())
-            } else {
-                it.toString()
-            } }
-    }
-}
 
-@KordPreview
-val MhguHzvCommand: suspend PublicSlashCommand<out MhguHzv>.() -> Unit = {
+
+val MhguHzvCommand: suspend PublicSlashCommand<out MhguHzv, ModalForm>.() -> Unit = {
     name = "hzv"
     description = "Get hzv info for a specific monster\n"
 
@@ -41,8 +32,8 @@ val MhguHzvCommand: suspend PublicSlashCommand<out MhguHzv>.() -> Unit = {
                 val icon = File("$dir/monster_icons/${monsterName}.webp")
                 val hzv = File("$dir/monster_hzv/${monsterName}.PNG")
 
-                files.add(NamedFile("icon.webp", icon.inputStream()))
-                files.add(NamedFile("hzv.PNG", hzv.inputStream()))
+                files.add(NamedFile("icon.webp", ChannelProvider { icon.inputStream().toByteReadChannel() }))
+                files.add(NamedFile("hzv.PNG", ChannelProvider { hzv.inputStream().toByteReadChannel() }))
 
                 embed {
                     title = "__${monsterName}__"