+++ src/main/kotlin/extensions/support.kt
@@ -4,14 +4,12 @@ import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
 import com.kotlindiscord.kord.extensions.types.respond
 import dev.kord.common.Color
-import dev.kord.common.annotation.KordPreview
-import dev.kord.common.entity.Snowflake
 import dev.kord.rest.builder.message.create.embed
+import env
 import kotlinx.datetime.Clock
 import utils.CatBot
 import java.util.logging.Logger
 
-@OptIn(KordPreview::class)
 class Support: Extension() {
 
     companion object {
@@ -25,7 +23,7 @@ class Support: Extension() {
             name = "support"
             description = "Support server for CatBot"
 
-            val invite = "https://discord.gg/FugAFKzTMw"
+            val invite = env.get("invite_url")
             action {
 
                 LOG.info("Received command: support")
@@ -38,7 +36,7 @@ class Support: Extension() {
                         description = "[Join the Discord]($invite)"
 
                         timestamp = Clock.System.now()
-                        footer { text = "Support Link Request"; icon = kord.getSelf().avatar?.url  }
+                        footer { text = "Support Link Request"; icon = kord.getSelf().avatar?.cdnUrl?.toUrl()  }
                     }
                 }
             }