+++ src/main/kotlin/main.kt
@@ -1,21 +1,28 @@
 import com.kotlindiscord.kord.extensions.ExtensibleBot
-import dev.kord.common.annotation.KordPreview
+import dev.kord.common.Color
 import dev.kord.gateway.Intents
-import dev.kord.gateway.PrivilegedIntent
+import dev.kord.gateway.NON_PRIVILEGED
+import dev.kord.rest.builder.message.create.embed
 import extensions.*
 import io.github.cdimascio.dotenv.dotenv
+import kotlinx.datetime.Clock
+import utils.CatBot
+import utils.CatBotError
 import java.util.logging.Logger
 
 val env = dotenv()
 
-@OptIn(PrivilegedIntent::class, KordPreview::class)
 suspend fun main() {
-
     val log = Logger.getLogger("Main")
 
+    var avatarUrl: String? = null;
     val client = ExtensibleBot(env["bot_token"]) {
         intents {
-            +Intents.nonPrivileged
+            +Intents.NON_PRIVILEGED
+        }
+
+        presence {
+            playing("Type / and click CatBot to get started")
         }
 
         extensions {
@@ -28,6 +35,23 @@ suspend fun main() {
             add( ::Mhr )
             add( ::Mhw )
             add( ::Support )
+        }
+
+        errorResponse { message, type ->
+            val invite = env.get("invite_url")
+            embed {
+                color = Color.CatBotError
+                title = "CatBot Error"
+                description = "Looks like we couldn't handle that request right now, try again later. if this keeps on happening [Join the Discord]($invite) for support"
+
+                timestamp = Clock.System.now()
+                footer {
+                    text = "Support Link Request";
+                    if (avatarUrl != null) {
+                        icon = avatarUrl
+                    }
+                }
+            }
         }
 
         hooks {
@@ -51,5 +75,6 @@ suspend fun main() {
         }
     }
 
+    avatarUrl = client.kordRef.getSelf().avatar?.cdnUrl?.toUrl()
     client.start()
 }