+++ src/main/kotlin/serializers/MHWWeaponResponse.kt
@@ -1,20 +1,8 @@
 package serializers
 
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.builtins.nullable
 import kotlinx.serialization.builtins.serializer
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.JsonTransformingSerializer
 
-// Little Sketchy, We trick the compiler into allowing us to pass in a Nullable Serializer
-// but seems to work, regardless :D
-object EmptyInt : JsonTransformingSerializer<Int>(Int.serializer().nullable as KSerializer<Int>) {
-    override fun transformDeserialize(element: JsonElement): JsonElement {
-        return if (element == JsonPrimitive("-")) JsonPrimitive(null as Number?) else element
-    }
-}
 
 @Serializable
 data class WeaponSharpness (
@@ -77,8 +65,9 @@ object WeaponMaterialSerializer : EmptyListSerializer<WeaponMaterial>(WeaponMate
 object WeaponElementsSerializer : EmptyListSerializer<WeaponElement>(WeaponElement.serializer())
 object WeaponAmmosSerializer : EmptyListSerializer<WeaponAmmo>(WeaponAmmo.serializer())
 
-class NullableWeaponShelling: NullableSerializer<WeaponShelling?>("NullableWeaponShelling", WeaponShelling.serializer().nullable)
-class NullableWeaponSharpness: NullableSerializer<WeaponSharpness?>("NullableWeaponSharpness", WeaponSharpness.serializer().nullable)
+object NullableWeaponShelling: NullableSerializer<WeaponShelling>(WeaponShelling.serializer())
+object NullableWeaponSharpness: NullableSerializer<WeaponSharpness>(WeaponSharpness.serializer())
+
 
 @Serializable
 data class MHWWeaponResponse(