+++ src/main/kotlin/utils/EmbedColor.kt
@@ -4,3 +4,6 @@ import dev.kord.common.Color
 
 val Color.Companion.CatBot: Color
     get() = Color(143, 222, 93)
+
+val Color.Companion.CatBotError: Color
+    get() = Color(255,0,0)